Move tracing helpers into a separate file
